requirements.txt
alembic
sqlalchemy
psycopg2-binary
supabase
python-dotenv
gradio
pandas
torch
transformers
peft
trl
datasets
accelerate
sentencepiece
huggingface-hub
